4. Формы. Загрузка файла на сервер. [#504855]: done

+++ src/main/java/ru/job4j/dreamjob/service/SimpleCandidateService.java
@@ -1,6 +1,7 @@
 package ru.job4j.dreamjob.service;
 
 import org.springframework.stereotype.Service;
+import ru.job4j.dreamjob.dto.FileDto;
 import ru.job4j.dreamjob.model.Candidate;
 import ru.job4j.dreamjob.repository.CandidateRepository;
 
@@ -16,23 +17,46 @@ public class SimpleCandidateService implements CandidateService {
 
 	private final CandidateRepository candidateRepository;
 
-	public SimpleCandidateService(CandidateRepository candidateRepository) {
+	private final FileService fileService;
+
+	public SimpleCandidateService(CandidateRepository candidateRepository, FileService fileService) {
 		this.candidateRepository = candidateRepository;
+		this.fileService = fileService;
 	}
 
 	@Override
-	public Candidate save(Candidate candidate) {
+	public Candidate save(Candidate candidate, FileDto fileDto) {
+		saveNewFile(candidate, fileDto);
 		return candidateRepository.save(candidate);
+	}
+
+	private void saveNewFile(Candidate candidate, FileDto fileDto) {
+		var file = fileService.save(fileDto);
+		candidate.setFileId(file.getId());
 	}
 
 	@Override
 	public Optional<Candidate> deleteById(int id) {
-		return candidateRepository.deleteById(id);
+		var fileoptional = findById(id);
+		Optional<Candidate> result = Optional.empty();
+		if (fileoptional.isPresent()) {
+			result = candidateRepository.deleteById(id);
+			fileService.deleteBiId(fileoptional.get().getFileId());
+		}
+		return result;
 	}
 
 	@Override
-	public boolean update(Candidate candidate) {
-		return candidateRepository.update(candidate);
+	public boolean update(Candidate candidate, FileDto fileDto) {
+		boolean isNewFileEmpty = fileDto.getContent().length == 0;
+		if (isNewFileEmpty) {
+			return candidateRepository.update(candidate);
+		}
+		int oldFileId = candidate.getFileId();
+		saveNewFile(candidate, fileDto);
+		boolean isUpdated = candidateRepository.update(candidate);
+		fileService.deleteBiId(oldFileId);
+		return isUpdated;
 	}
 
 	@Override
